requirements.txt
hear21passt
pandas
pytorch-lightning
torchaudio
sed_eval
scikit-learn
soundfile
librosa
